Abstract engine service and add Unsloth support
- Refactored `MLXEngineService` into `BaseEngineService` and `MLXEngineService`.
- Added `UnslothEngineService` with support for Unsloth/CUDA fine-tuning.
- Implemented `EngineFactory` to handle engine selection and hardware detection.
- Updated API to use the factory and expose engine status/selection endpoints.
- Added frontend `EngineSelectionModal` to allow users to choose the engine.
- Updated `pyproject.toml` with Unsloth dependencies.
- Added import guards for Unsloth to allow running on non-CUDA hardware (fallback).
